Do all the work in a separate logical thread
We should be responding to the incoming http request from slack
immediately, to avoid the scenario where we take too long and
slack sends it again.

Moved the effects to read env variables closer to where they are
consumed. When I set this up in a new space, you need to get some
keys from Slack and you need to e.g. reply to the `url_verification`
before you can get some of the keys.